Update UTests, also fix some functionality with incapsulation
